+++ CHANGELOG.md
@@ -24,8 +24,11 @@
 
 - updated bar chart event handlers to adapt to the changes in `@svizzle/barchart` v0.11.0
 
-## `@svizzle/ui` v0.10.1 (next)
+## `@svizzle/ui` v0.11.0 (next)
 
+- `Scroller`: exposed 2 new props
+	- `extraWidth`, to account for scrollbar width on Windows
+	- `outerScrollTop`, to let other components binding to its internal scroll
 - `ScreenSensor`: fixed the measured `glyph.height` being too big
 - `Banner`, `Switch` and `XorSelector`: resolved some a11y related warnings